Client Updates
Redesigned the Ini reader to read everything into runtime variables. All of the game's settings will be loaded and stored using this. Ini is outdated but it should be easier to port the game to other OSs like this. (I don't like XML)

Rearranged some other stuff and added a few more ? chat commands.